+++ src/main/java/org/folio/marc4ld/service/marc2ld/bib/mapper/additional/DictionaryBasedMarc2LdMapper.java
@@ -1,6 +1,9 @@
 package org.folio.marc4ld.service.marc2ld.bib.mapper.additional;
 
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
 import org.folio.ld.dictionary.PropertyDictionary;
 import org.folio.ld.dictionary.model.Resource;
 import org.folio.marc4ld.dto.MarcData;
@@ -20,18 +23,25 @@ public abstract class DictionaryBasedMarc2LdMapper implements AdditionalMapper {
 
   @Override
   public void map(MarcData marcData, Resource resource) {
-    MappingConfig config = getMappingConfig();
+    getMappingConfigs()
+      .stream()
+      .map(mc -> mapSingle(resource, mc))
+      .filter(Objects::nonNull)
+      .forEach(properties -> resource.setDoc(mapperHelper.getJsonNode(properties)));
+  }
+
+  private Map<String, List<String>> mapSingle(Resource resource, MappingConfig config) {
     var properties = mapperHelper.getProperties(resource);
     var codes = properties.getOrDefault(config.keyProperty.getValue(), List.of());
     var values = dictionaryProcessor.getValues(config.dictionaryName, codes);
     if (values.isEmpty()) {
-      return;
+      return null;
     }
     properties.put(config.valueProperty.getValue(), values);
-    resource.setDoc(mapperHelper.getJsonNode(properties));
+    return properties;
   }
 
-  abstract MappingConfig getMappingConfig();
+  abstract Set<MappingConfig> getMappingConfigs();
 
   record MappingConfig(String dictionaryName, PropertyDictionary keyProperty, PropertyDictionary valueProperty) {
   }